+++ src/components/index.js
@@ -18,3 +18,5 @@ export { default as WarningMessage } from './WarningMessage';
 export { default as CalloutRenderer } from './CalloutRenderer';
 export { default as ViewSource } from './ViewSource';
 export { default as AdministrativeNoteList } from './AdministrativeNoteList';
+export { default as MultiSelectionFacet } from './MultiSelectionFacet';
+export { default as FacetOptionFormatter } from './FacetOptionFormatter';