+++ CHANGELOG.md
@@ -52,6 +52,7 @@ Inventory causes an error. Refs UIIN-2012.
 * Fix option View Source is not available in holdings record view. Refs UIIN-2044.
 * Browse contributors list. Refs UIIN-2024.
 * Browse form - first and last record navigation. Refs UIIN-1912.
+* Inventory >  Contributors Browse - Create a Name Type facet. Refs UIIN-2021.
 
 ## [9.0.0](https://github.com/folio-org/ui-inventory/tree/v9.0.0) (2022-03-03)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v8.0.0...v9.0.0)